tests/wpt/tests/webrtc-extensions/RTCOAuthCredential.html

<!doctype html>
<meta charset=utf-8>
<title>RTCConfiguration iceServers with OAuth credentials</title>
<script src='/resources/testharness.js'></script>
<script src='/resources/testharnessreport.js'></script>
<script src='../webrtc/RTCConfiguration-helper.js'></script>
<script>
  'use strict';

// These tests are based on
// https://w3c.github.io/webrtc-extensions/#rtcoauthcredential-dictionary

/*
    4.3.2.  To set a configuration
    11.6. If scheme name is turn or turns, and server.credentialType is "oauth",
    and server.credential is not an RTCOAuthCredential, then throw an
    InvalidAccessError and abort these steps.
*/
config_test(makePc => {
  assert_throws_dom('InvalidAccessError', () =>
    makePc({ iceServers: [{
      urls: 'turns:turn.example.org',
      credentialType: 'oauth',
      username: 'user',
      credential: 'cred'
    }] }));
}, 'with turns server, credentialType oauth, and string credential should throw InvalidAccessError');

config_test(makePc => {
  const pc = makePc({ iceServers: [{
    urls: 'turns:turn2.example.net',
    username: '22BIjxU93h/IgwEb',
    credential: {
      macKey: 'WmtzanB3ZW9peFhtdm42NzUzNG0=',
      accessToken: 'AAwg3kPHWPfvk9bDFL936wYvkoctMADzQ5VhNDgeMR3+ZlZ35byg972fW8QjpEl7bx91YLBPFsIhsxloWcXPhA=='
    },
    credentialType: 'oauth'
  }]});
  const { iceServers } = pc.getConfiguration();
  const server = iceServers[0];
  assert_equals(server.credentialType, 'oauth');
}, 'with turns server, credential type and credential from spec should not throw');

</script>
