third_party/blink/web_tests/external/wpt/webrtc-extensions/RTCOAuthCredential-expected.txt

This is a testharness.js-based test.
FAIL new RTCPeerConnection(config) - with turns server, credentialType oauth, and string credential should throw InvalidAccessError assert_throws_dom: function "() =>
    makePc({ iceServers: [{
      urls: 'turns:turn.example.org',
      credentialType: 'oauth',
      username: 'user',
      credential: 'cred'
    }] })" did not throw
FAIL setConfiguration(config) - with turns server, credentialType oauth, and string credential should throw InvalidAccessError assert_throws_dom: function "() =>
    makePc({ iceServers: [{
      urls: 'turns:turn.example.org',
      credentialType: 'oauth',
      username: 'user',
      credential: 'cred'
    }] })" did not throw
FAIL new RTCPeerConnection(config) - with turns server, credential type and credential from spec should not throw assert_equals: expected (string) "oauth" but got (undefined) undefined
FAIL setConfiguration(config) - with turns server, credential type and credential from spec should not throw assert_equals: expected (string) "oauth" but got (undefined) undefined
Harness: the test ran to completion.
